backend/requirements.txt
Django>=5.0,<6
 djangorestframework>=3.15
 gunicorn>=21.2
 uvicorn[standard]>=0.30
 psycopg2-binary>=2.9
 python-dotenv>=1.0
 whitenoise>=6.7
